Move handler404 to its own file for testing
